% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettewidget.R
\name{palettewidget}
\alias{palettewidget}
\title{Color Palette Widget}
\usage{
palettewidget(palette, displayNames = FALSE)
}
\arguments{
\item{palette}{The color palette to be displayed.}

\item{displayNames}{Use `TRUE` to attempt to display HTML/CSS color names
for those colors that have them.}
}
\value{
The rendered color palette widget.
}
\description{
Displays a color palette and their HTML/CSS hex color codes or names.
}
\examples{
palettewidget(c("#ff0000", "#00ff00", "#0000ff"))
palettewidget(c("#ff0000", "#00ff00", "#0000ff"),
              displayNames = TRUE)  # should display "red", "lime", "blue"

}

